# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vivd Catalyst API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/conversations": {
      "get": {
        "operationId": "listConversations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "clientInstanceId": {
                        "type": "string"
                      },
                      "ownerUserId": {
                        "type": "string"
                      },
                      "ownerExternalUserId": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      },
                      "retainedUntil": {
                        "type": "string"
                      },
                      "deletedAt": {
                        "type": "string"
                      },
                      "latestMessageAt": {
                        "type": "string"
                      },
                      "activeRun": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "conversationId": {
                            "type": "string"
                          },
                          "agentName": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "queued",
                              "running",
                              "waiting_for_permission",
                              "cancelling"
                            ]
                          },
                          "startedAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "lastSequence": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "id",
                          "conversationId",
                          "agentName",
                          "status",
                          "startedAt",
                          "updatedAt",
                          "lastSequence"
                        ],
                        "additionalProperties": false
                      },
                      "unread": {
                        "type": "boolean"
                      },
                      "lastViewedAt": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "clientInstanceId",
                      "ownerUserId",
                      "ownerExternalUserId",
                      "title",
                      "status",
                      "createdAt",
                      "updatedAt",
                      "retainedUntil"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createConversation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "locale": {
                    "type": "string",
                    "enum": [
                      "en",
                      "de"
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "clientInstanceId": {
                      "type": "string"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "ownerExternalUserId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "retainedUntil": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "clientInstanceId",
                    "ownerUserId",
                    "ownerExternalUserId",
                    "title",
                    "status",
                    "createdAt",
                    "updatedAt",
                    "retainedUntil"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{conversationId}/title": {
      "post": {
        "operationId": "generateConversationTitle",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "clientInstanceId": {
                      "type": "string"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "ownerExternalUserId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "retainedUntil": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "clientInstanceId",
                    "ownerUserId",
                    "ownerExternalUserId",
                    "title",
                    "status",
                    "createdAt",
                    "updatedAt",
                    "retainedUntil"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}